Group feed store specs and helpers
