+++ PokemonEncyclopedia/PokemonEncyclopediaTests/FeedCache/XCTestCase+FailableDeleteFeedStoreSpecs.swift
@@ -1,0 +1,26 @@
+//
+//  XCTestCase+FailableDeleteFeedStoreSpecs.swift
+//  PokemonEncyclopediaTests
+//
+//  Created by Ankit on 19/07/21.
+//
+
+import XCTest
+import PokemonEncyclopedia
+
+extension FailableDeleteFeedStoreSpecs where Self: XCTestCase {
+	
+	func assertThatDeleteDeliversErrorOnDeletionError(on sut: FeedStore, file: StaticString = #file, line: UInt = #line) {
+		let deletionError = deleteCache(from: sut)
+
+		XCTAssertNotNil(deletionError, "Expected cache deletion to fail", file: file, line: line)
+	}
+	
+	
+	func assertThatDeletehasNoSideEffectsOnDeletionError(on sut: FeedStore, file: StaticString = #file, line: UInt = #line) {
+		deleteCache(from: sut)
+
+		expect(sut, toRetrieve: .empty, file: file, line: line)
+	}
+	
+}
