+++ PokemonEncyclopedia/PokemonEncyclopediaTests/FeedCache/XCTestCase+FeedStoreSpecs.swift
@@ -151,63 +151,4 @@ extension FeedStoreSpecs where Self: XCTestCase {
 		expect(sut, toRetrieve: expectedResult, file: file, line: line)
 	}
 	
-	
-}
-
-
-extension FailableRetrieveFeedStoreSpecs where Self: XCTestCase {
-	
-	func assertThatRetrieveDeliversFailureOnRetrievalError(on sut: FeedStore, file: StaticString = #file, line: UInt = #line) {
-		
-		expect(sut, toRetrieve: .failure(anyNSError()), file: file, line: line)
-	}
-	
-	func assertThatRetrieveHasNoSideEffectsOnFailure(on sut: FeedStore, file: StaticString = #file, line: UInt = #line) {
-		
-		expect(sut, toRetrieveTwice: .failure(anyNSError()), file: file, line: line)
-	}
-	
-}
-
-
-
-extension FailableInsertFeedStoreSpecs where Self: XCTestCase {
-	
-	func assertThatInsertDeliversErrorOnInsertionError(on sut: FeedStore, file: StaticString = #file, line: UInt = #line) {
-		let feed = uniqueImageFeed().local
-		let timestamp = Date()
-		
-		let insertionError = insert((feed, timestamp), to: sut)
-		
-		XCTAssertNotNil(insertionError, "Expect cache insertion error", file: file, line: line)
-	}
-	
-	
-	func assertThatInsertHasNoSideEffectsOnInsertionError(on sut: FeedStore, file: StaticString = #file, line: UInt = #line) {
-		let feed = uniqueImageFeed().local
-		let timestamp = Date()
-		
-		insert((feed, timestamp), to: sut)
-		
-		expect(sut, toRetrieve: .empty, file: file, line: line)
-	}
-	
-	
-}
-
-extension FailableDeleteFeedStoreSpecs where Self: XCTestCase {
-	
-	func assertThatDeleteDeliversErrorOnDeletionError(on sut: FeedStore, file: StaticString = #file, line: UInt = #line) {
-		let deletionError = deleteCache(from: sut)
-
-		XCTAssertNotNil(deletionError, "Expected cache deletion to fail", file: file, line: line)
-	}
-	
-	
-	func assertThatDeletehasNoSideEffectsOnDeletionError(on sut: FeedStore, file: StaticString = #file, line: UInt = #line) {
-		deleteCache(from: sut)
-
-		expect(sut, toRetrieve: .empty, file: file, line: line)
-	}
-	
 }
